fix: Remove managedFields from AppInstance (#397)
The `kubit helpers fetch-app-instance` command is used to populate
`/overlay/appinstance.json` during the apply job. This command fetches
the current state of the AppInstance with the API but does not strip out
managedFields. This leads to kubectl returning
`Error from server (BadRequest): metadata.managedFields must be nil`
when trying to apply the manifests.

Co-authored-by: Jack <56563911+jdockerty@users.noreply.github.com>